BE/requirements.txt
# Core
python-dateutil>=2.8.2
pydantic>=2.0.0
pydantic-settings>=2.0.0

# RSS Parsing
feedparser>=6.0.10

# HTTP Client
aiohttp>=3.9.0
httpx>=0.25.0

# AI Processing
google-genai

# Data Processing
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
